% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btloda.R
\name{btloda}
\alias{btloda}
\title{This function returns the anomaly scores, the projections and histograms built from using bootstrap and out-of-bag version of loda}
\usage{
btloda(a, sparsity = NA, maxk = 1000, inf_replace = log(1e-09))
}
\arguments{
\item{a}{Input dataframe or data matrix}

\item{sparsity}{The proportion of the dimensions assigned to be 0 when build each random projection. Default is NA, in which case if the data dimension d > 1, 1 - d^{\{-1/2\}} proportion of dimensions will be set to 0, and if d = 1, no dimension will be set to 0.}

\item{maxk}{The number of projections/histograms we want to build.}

\item{inf_replace}{The quantity to use to replace the possible negative infinity got from calculating log density based on the histograms built. Default is log(1e-09).}
}
\value{
A list consisting of the anomaly scores, the projections and histograms built.
}
\description{
This function returns the anomaly scores, the projections and histograms built from using bootstrap and out-of-bag version of loda
}
\examples{
a = matrix(rnorm(200, 0, 1), nrow = 20, ncol = 10)
bt_out = btloda(a,sparsity=NA, maxk=1000, inf_replace = log(1e-09))
bt_out$oob_nll 
bt_out$pvh$w
bt_out$pvh$hists
}
