% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oob_get_neg_ll_all_hist.R
\name{oob_get_neg_ll_all_hist}
\alias{oob_get_neg_ll_all_hist}
\title{get out-of-bag negative loglikelihood from all histograms that didn't use the point}
\usage{
oob_get_neg_ll_all_hist(a, record_mat, w, hists, inf_replace = NA)
}
\description{
get out-of-bag negative loglikelihood from all histograms that didn't use the point
}
