% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btloda.R
\name{btloda}
\alias{btloda}
\title{This function returns both the anomaly scores and the projection results}
\usage{
btloda(a, sparsity = NA, maxk = 1000, keep = NULL, exclude = NULL,
  debug = F, inf_replace = log(1e-09))
}
\description{
This function returns both the anomaly scores and the projection results
}
