% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_neg_ll_all_hist.R
\name{get_neg_ll_all_hist}
\alias{get_neg_ll_all_hist}
\title{get negative loglikelihood from all histograms}
\usage{
get_neg_ll_all_hist(a, w, hists, inf_replace = NA)
}
\description{
get negative loglikelihood from all histograms
}
